Added clicker support for Flowtime
For those using presenter devices, like the Logitech R400, the normal key mappings
in Flowtime don't make sense because the next and prev buttons really just jump
to the top and bottom of sections (which isn't very useful).

This change adds a function to the public API called .  When set to true,
the bindings are changed so that keycodes 33 and 34 are used for slide navigation
(instead of top of section and bottom of section).

This change is completely backwards compatible and the default is to keep the
old behavior.  Enabling the clicker support really doesn't change much anyway because
I never used the previous functions of keycodes 33 and 34 anyway.